+++ gradle-plugin/src/integTest/java/com/github/dynamicextensionsalfresco/gradle/integrationTest/AbstractIntegrationTest.java
@@ -26,9 +26,10 @@ public abstract class AbstractIntegrationTest {
    @Parameters(name = "Gradle v{0}")
    public static Collection<Object[]> testData() {
        return Arrays.asList(new Object[][]{
-               {"6.7.1"},
+               {"6.8.3"},
+               {"6.0"},
                {"5.6.4"},
-               {"4.10.3"},
+               {"5.3"},
        });
     }
 
@@ -45,7 +46,6 @@ public abstract class AbstractIntegrationTest {
                 .withPluginClasspath()
                 .withGradleVersion(gradleVersion)
                 .withArguments(task, "--stacktrace", "--rerun-tasks")
-                .withDebug(true)
                 .forwardOutput()
                 .build();
     }