+++ gradle-plugin/src/main/java/com/github/dynamicextensionsalfresco/gradle/tasks/DeBundleTaskConvention.java
@@ -2,7 +2,7 @@ package com.github.dynamicextensionsalfresco.gradle.tasks;
 
 import aQute.bnd.gradle.BundleTaskConvention;
 import aQute.bnd.osgi.Constants;
-import com.github.dynamicextensionsalfresco.gradle.internal.BndHandler;
+import java.util.Arrays;
 import org.gradle.api.Action;
 import org.gradle.api.Task;
 import org.gradle.api.tasks.bundling.Jar;
@@ -13,7 +13,6 @@ import org.gradle.api.tasks.bundling.Jar;
  * The TaskConvention does not contribute any additional user-modifiable properties or methods to the task it is applied to.
  */
 public class DeBundleTaskConvention {
-    private final Jar task;
     private final BundleTaskConvention bndConvention;
 
     /**
@@ -22,7 +21,6 @@ public class DeBundleTaskConvention {
      * @param task The {@link Jar} task the convention will be applied to
      */
     public DeBundleTaskConvention(Jar task) {
-        this.task = task;
         bndConvention = new BundleTaskConvention(task);
         task.getConvention().getPlugins().put("_bundle_bnd", bndConvention);
         // This is an anonymous inner class instead of a lambda because of a gradle issue where lambdas break up-to-date checks
@@ -34,18 +32,33 @@ public class DeBundleTaskConvention {
         });
     }
 
-    /**
-     * Internal function that is called in a {@link org.gradle.api.Task#doLast(Action)} action to update OSGi headers and then run the BND bundle build
-     */
     private void buildDeBundle() {
-        BndHandler bndHandler = new BndHandler(task, bndConvention);
+        bndConvention.bnd("Alfresco-Dynamic-Extension=true");
 
-        bndHandler.setHeader("Alfresco-Dynamic-Extension", "true");
-        if(!bndHandler.hasHeader(Constants.IMPORT_PACKAGE)) {
-            bndHandler.setHeader(Constants.IMPORT_PACKAGE, "*");
-            bndHandler.setHeader(Constants.DYNAMICIMPORT_PACKAGE, "*");
+        boolean hasImportPackage = Arrays.stream(bndConvention.getBnd().get().split("\n")).anyMatch(s -> s.startsWith(Constants.IMPORT_PACKAGE+"="));
+        if(!hasImportPackage) {
+            bndConvention.bnd(
+                    Constants.IMPORT_PACKAGE+"=*",
+                    Constants.DYNAMICIMPORT_PACKAGE+"=*"
+            );
         }
         bndConvention.buildBundle();
     }
 
+    public BundleTaskConvention getBndConvention() {
+        return bndConvention;
+    }
+
+    public static DeBundleTaskConvention apply(Jar jar) {
+        DeBundleTaskConvention convention = jar.getConvention().findPlugin(DeBundleTaskConvention.class);
+        if(convention == null) {
+            convention = new DeBundleTaskConvention(jar);
+            jar.getConvention().getPlugins().put("bundle", convention);
+        }
+        return convention;
+    }
+
+    public static DeBundleTaskConvention get(Jar jar) {
+        return jar.getConvention().getPlugin(DeBundleTaskConvention.class);
+    }
 }